requirements.txt
Flask==3.0.3
gunicorn==21.2.0

numpy>=1.26.0,<3.0
pandas==2.2.3
scikit-learn==1.6.0
joblib>=1.3.0

tensorflow-cpu==2.16.2

matplotlib==3.7.1
typing_extensions>=4.8.0
itsdangerous>=2.1.0
Werkzeug>=3.0.0
Jinja2>=3.1.0
